Rename 1234punch.ino.ino to 1234punch.ino
My project is a simple animation which integrates a button switch built in class using foam, copper, connecting cables from our starter kit and a USB-C cable connecting the arduino kit to the computer, taking in commands. Within the arduino code, I simply used a button command and had the code alternate between 2 different animations. When the user presses the practically built switch I made,  the character appears to be punching the edges of the screen, almost like they are trying to break out of the display. Even when the switch isn’t being pressed, the character is constantly moving and appearing full of life.